calculator modules

+++ 0x02-python-import_modules/calculator_1.py
@@ -12,4 +12,4 @@ def mult(a, b):
 
 
 def div(a, b):
-    return a / b
+    return int(a / b)
